Collecting items after the pve round is now possible
Implemented a system for collecting orbs (gray, blue and gold) after the pve round.
Added a picture with "question mark" to find it on the board.
The number of item searches has been set to 3. Maybe in the future we should find another method of determining how many items there are on the map.
Two coordinates have been added. For the place where the orbs fall. And the character's starting position to get back there after your search.